Move Recorder to Basic implementation
